Add Hostname Support while on SSH
