+++ mods/directories.go
@@ -3,7 +3,6 @@ package mods
 import (
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 
 	"github.com/talal/go-bits/color"
@@ -30,6 +29,7 @@ func getDir(cwd string) string {
 	gbpath := pathToDisplay[strings.LastIndex(pathToDisplay, "/")+1:]
 	gitDir, err := findGitRepo(cwd)
 	handleError(err)
+
 	isconmod := iscontentmodified(gitDir)
 
 	if gitDir != "" {
@@ -38,7 +38,7 @@ func getDir(cwd string) string {
 		imod := <-status
 		return imod
 	}
-	return color.Sprintf(color.Cyan, pathToDisplay)
+	return color.Sprintf(color.BrightCyan, pathToDisplay)
 }
 
 func findNearestAccessiblePath(path string) string {
@@ -48,31 +48,4 @@ func findNearestAccessiblePath(path string) string {
 	}
 
 	return findNearestAccessiblePath(filepath.Dir(path))
-}
-
-func findstatus(mods ismodified, path string, gdir string, status chan string) {
-	branchchannel := make(chan string)
-	go currentGitBranch(gdir, branchchannel)
-	branch := <-branchchannel
-	nm := color.Sprintf(color.BrightYellow, path) + " on " +
-		color.Sprintf(color.BrightGreen, ` `+branch)
-	if mods.notStaged != 0 && mods.untracked != 0 {
-		status <- nm + color.Sprintf(color.BrightRed, ` [`+strconv.Itoa(mods.notStaged)+`!]`+`[`+strconv.Itoa(mods.untracked)+`+]`)
-	}
-	if mods.notStaged != 0 {
-		status <- nm + color.Sprintf(color.BrightRed, ` [`+strconv.Itoa(mods.notStaged)+`!]`)
-	}
-	if mods.untracked != 0 {
-		status <- nm + color.Sprintf(color.BrightRed, ` [`+strconv.Itoa(mods.untracked)+`+]`)
-	}
-	cmod := make(chan int)
-
-	go isahead(path, branch, cmod)
-	ahead := <-cmod
-	//fmt.Print(ahead)
-	if ahead != 0 {
-		status <- nm + color.Sprintf(color.BrightRed, ` [`+strconv.Itoa(ahead)+`x]`)
-	}
-
-	status <- nm
 }
