+++ mods/fns.go
@@ -9,6 +9,8 @@ import (
 	"github.com/talal/go-bits/color"
 )
 
+var host = gethost()
+
 func cwdir() string {
 	wd, err := os.Getwd()
 	if err != nil {
@@ -23,11 +25,18 @@ func emptifier(list []string, val string) []string {
 	return append(list, val)
 }
 func stripHomeDir(path string) string {
+	ssh := make(chan bool)
+	go isssh(ssh)
+	sshtrue := <-ssh
+	host := gethost()
+	if sshtrue == true {
+		return strings.Replace(path, os.Getenv("HOME"), color.Sprintf(color.BrightGreen, host)+" ~", 1)
+	}
 	return strings.Replace(path, os.Getenv("HOME"), "~", 1)
 }
 func handleError(err error) {
 	if err != nil {
-		color.Fprintf(os.Stderr, color.Red, "Prompt error: %v\n", err)
+		color.Fprintf(os.Stderr, color.BrightRed, "Prompt error: %v\n", err)
 	}
 }
 func getExitCodeField(arg string) string {
